Properly override dev-lib commands on Travis
We need to place this in this function that won’t be called until the
commands are ready to be overridden, because this project config file
is loaded twice, once in `.travis.yml` (to allow it to override the
dev-lib path), and again in `travis/before_script.sh`. This causes our
builds to error with:

```
dev-lib/travis/commands.sh: line 42: setup-phpunit: command not found
```

When the before-script file is sourced, because the `setup-phpunit`
alias is already set up before `commands.sh` is loaded.

See #10