Fix issues with the activity events
While working on #17, several issues were discovered in the Activity Update Post and Activity Comment Post hook events.

- The ham action was specifying the arg index incorrectly, the activity is the first parameter, not the third. Because of the way that the unit tests were constrcted, they did not detect this. They have now been updated.
- The ham and spam were firing for any activity type, not just the one in question (update or comment, respectively). This has been fixed by introducing an action object that checks the activity type, and the tests have been updated to check for this as well. This isn't such a big deal for spam, because of the way that the Points reactor handles reversals. However, for ham it is a big issue.

Issues were also discovered with the autoloader classmap, that we weren't aware of until now. Getting everything working correctly in regard to that also required a few changes to the dev-lib.

See #17, #16, #4