Fix failures in group image upload tests
Both the cover image and avatar tests for groups related events were failing. For the avatars, this was only the tests for the `current:user` target, and this was caused by the current user arg not retrieving the value correctly when the testcase was trying to run the assertion for some reason. The cover image tests also had an issue with the current group being set by prior tests, causing the upload to be related to the wrong group.

See #6