Fix tests exiting on WP 4.6
The `'wp_die_ajax_handler'` action is being used here to prevent `wp_die()` from actually exiting. However, that filter is only used when doing Ajax, but we only tell WordPress that we are doing Ajax via the `'wp_doing_ajax'` filter, which wasn't introduced until WordPress 4.7. So we need to also hook into the `'wp_die_handler'` action as a fallback on 4.6.

See #5, #6