Avatar upload event: only fire when an avatar is set
And only reverse when it is deleted, not just when it is replaced with
a new one.

Previously, uploading a new avatar would remove points for the previous
one and then add points for the new one, generating an unnecessary set
of points transactions with a sum of zero.

See #39