Group avatar set: only award points on initial upload
The points were being awarded each time a new avatar was uploaded, and
then removed all at once when the avatar was actually deleted instead of
just replaced. Instead, the behavior is now corrected to award only when
setting the avatar, not replacing it, and to delete only on delete.

Fixes #42