Groups cover image set: only award points on initial upload
Not each time that the cover image is replaced with a new one.
Previously there was a bug that caused the points to be awarded each
time, and then removed all at once when the delete button is used. This
is now corrected so that points are only awarded when the cover image is
initially set, not when it is replaced, and only removed when it is
deleted completely.

Fixes #43